refactor profile uploader
