+++ lib/saito/ui/photo-uploader/photo-uploader.template.js
