+++ web/saito/css-imports/saito-photo-uploader.css
@@ -1,0 +1,46 @@
+
+.uploader-uploaded-photo {
+	height: auto;
+	object-fit: cover;
+	height: 100%;
+	width: 100%;
+	left: 0;
+	bottom: 0;
+	padding: 1rem;
+	display: none;
+	position: absolute;
+	height: 90%;
+	top: 0;
+}
+
+.photo-uploader {
+	background-color: var(--saito-background-color);
+	height: 60rem;
+	width: 70rem;
+	/* position: absolute; */
+	/* top: 50%;
+	left: 50%; */
+	padding: 5rem;
+	/* transform: translate(-50%, -50%); */
+	text-align: center;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	border: 1px solid;
+	font-size: 2rem;
+	border-radius: 0.5rem;
+	color: var(--saito-font-color);
+}
+
+.uploader-actions {
+	width: 100%;
+	position: absolute;
+	bottom: 0;
+	display: none;
+	margin-bottom: 1rem;
+	z-index: 10;
+}
+.uploader-actions > :nth-child(1) {
+	margin-right: 1rem;
+}
